fix: embed remote catalog config into remote runtime config
The `RemoteCatalogConfig` type contains fields necessary for
editing the configuration of a running remote MCP server. Today, the
configuration constraints are always taken directly from the catalog
entry, which can change post-launch. This means that when a catalog is
updated, the running remote MCP servers referencing it are automatically
updated, bypassing the normal upgrade flow for MCP servers.

This was also an issue for changes to composite MCP catalog entries.

As a part of fixing this, update the type definition of
`RemoteRuntimeConfig` so that it embeds `RemoteCatalogConfig`.

Note: In addition to this change, we'll need to implement some migration
logic to populate the fields from the current catalog entries where
possible.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>